requirements.txt
allennlp==2.10.1
Flask==3.0.3
numpy==1.24.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
overrides==3.0.0
paddleocr==2.8.1
paddlepaddle
pandas==2.0.3
Pillow==10.2.0
python_Levenshtein==0.25.1
scipy==1.10.1
tabulate==0.9.0
torch==1.12.1
torchsummary==1.5.1
torchtext==0.6.0
torchvision==0.13.1
tqdm==4.66.4
vietocr==0.3.13
Werkzeug==3.0.4
transformers==4.44.2
spacy==3.3.3
pydantic==1.10.2